Research-AI-main/backend/requirements.txt
flask
flask-cors
python-dotenv
groq
pypdf
requests
flask-sqlalchemy
flask-jwt-extended
flask-bcrypt
